+++ packages/react/src/components/ProductUpdates/src/components/ProductUpdatesFooter/index.ts
@@ -1,0 +1,5 @@
+import { ProductUpdatesFooter } from "./ProductUpdatesFooter";
+
+export { ProductUpdatesFooter } from "./ProductUpdatesFooter";
+export type { ProductUpdatesFooterProps } from "./productUpdatesFooter.types";
+export default ProductUpdatesFooter;