+++ packages/react/src/components/ProductUpdates/src/productUpdates.types.ts
@@ -1,5 +1,12 @@
 import { HTMLAttributes, ReactNode } from "react";
-import { PopoverProperties } from "@nimbus-ds/components";
+import { BoxProperties, PopoverProperties } from "@nimbus-ds/components";
+
+import { ProductUpdatesPaginator, ProductUpdatesFooter } from "./components";
+
+export interface ProductUpdatesComponents {
+  Paginator: typeof ProductUpdatesPaginator;
+  Footer: typeof ProductUpdatesFooter;
+}
 
 export interface ProductUpdatesProperties
   extends Omit<
@@ -15,10 +22,20 @@ export interface ProductUpdatesProperties
    */
   text: string;
   /**
+   * Optional tag element rendered inline before the title.
+   * @TJS-type React.ReactNode
+   */
+  tag?: ReactNode;
+  /**
    * Optional content for the body of the Popover.
    * @TJS-type React.ReactNode
    */
   bodyContent?: ReactNode;
+
+  /**
+   * Optional properties for the content container of the popover body.
+   */
+  bodyContentProps?: Omit<BoxProperties, "children">;
   /**
    * Optional bottom link to dismiss the Popover. If unset, an IconButton with an X will appear on the superior right corner.
    * @TJS-type React.ReactNode