+++ packages/react/src/components/ProductUpdates/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ProductUpdates component allows the user to showcase new features or updates in a non-distractive way by displaying the information in a floating popover with a high-contrast setting that catches the eye of the user.
 
+## 2026-03-02 `1.6.0`
+
+### 🎉 New features
+
+- Added `ProductUpdates.Paginator` and `ProductUpdates.Footer` as compound sub-components. ([#155](https://github.com/TiendaNube/nimbus-patterns/pull/155) by [@joacotornello](https://github.com/joacotornello))
+- Added `tag` prop to render an optional element inline before the title. ([#155](https://github.com/TiendaNube/nimbus-patterns/pull/155) by [@joacotornello](https://github.com/joacotornello))
+- Added `bodyContentProps` prop to pass extra properties to the popover body container. ([#155](https://github.com/TiendaNube/nimbus-patterns/pull/155) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2026-01-15 `1.5.1`
 
 #### 📚 3rd party library updates